Conflicts:
	dtgov-ui-war/src/main/java/org/overlord/dtgov/ui/client/local/beans/UiConfiguration.java
	dtgov-ui-war/src/main/java/org/overlord/dtgov/ui/client/local/site/dashboard.html
	dtgov-ui-war/src/main/java/org/overlord/dtgov/ui/server/services/WorkflowQueryService.java

+++ dtgov-ui-war/src/main/java/org/overlord/dtgov/ui/server/services/TargetService.java
@@ -18,9 +18,12 @@ package org.overlord.dtgov.ui.server.services;
 import java.util.List;
 
 import javax.inject.Inject;
+import javax.servlet.ServletRequest;
+import javax.servlet.http.HttpServletRequest;
 
 import org.apache.commons.lang.StringUtils;
 import org.jboss.errai.bus.server.annotations.Service;
+import org.jboss.errai.bus.server.api.RpcContext;
 import org.oasis_open.docs.s_ramp.ns.s_ramp_v1.BaseArtifactType;
 import org.overlord.dtgov.common.targets.TargetConstants;
 import org.overlord.dtgov.ui.client.shared.beans.TargetBean;
@@ -29,9 +32,11 @@ import org.overlord.dtgov.ui.client.shared.beans.ValidationError;
 import org.overlord.dtgov.ui.client.shared.exceptions.DtgovFormValidationException;
 import org.overlord.dtgov.ui.client.shared.exceptions.DtgovUiException;
 import org.overlord.dtgov.ui.client.shared.services.ITargetService;
+import org.overlord.dtgov.ui.server.i18n.Messages;
 import org.overlord.dtgov.ui.server.services.sramp.SrampApiClientAccessor;
 import org.overlord.dtgov.ui.server.services.targets.TargetFactory;
 import org.overlord.dtgov.ui.server.services.targets.TargetValidator;
+import org.overlord.dtgov.ui.server.util.AuthUtils;
 import org.overlord.sramp.atom.err.SrampAtomException;
 import org.overlord.sramp.client.SrampAtomApiClient;
 import org.overlord.sramp.client.SrampClientException;
@@ -53,15 +58,12 @@ public class TargetService implements ITargetService {
     @Inject
     private TargetValidator _targetValidator;
 
-    /*
-     * (non-Javadoc)
-     *
-     * @see
-     * org.overlord.dtgov.ui.client.shared.services.ITargetService#delete(java
-     * .lang.String)
+    /**
+     * @see org.overlord.dtgov.ui.client.shared.services.ITargetService#delete(java.lang.String)
      */
     @Override
     public void delete(String uuid) throws DtgovUiException {
+        checkAuthorization();
         try {
             _srampClientAccessor.getClient().deleteArtifact(uuid, ArtifactType.ExtendedArtifactType(TargetConstants.TARGET_EXTENDED_TYPE, false));
         } catch (SrampClientException e) {
@@ -71,15 +73,12 @@ public class TargetService implements ITargetService {
         }
     }
 
-    /*
-     * (non-Javadoc)
-     *
-     * @see
-     * org.overlord.dtgov.ui.client.shared.services.ITargetService#get(java.
-     * lang.String)
+    /**
+     * @see org.overlord.dtgov.ui.client.shared.services.ITargetService#get(java.lang.String)
      */
     @Override
     public TargetBean get(String uuid) throws DtgovUiException {
+        checkAuthorization();
         try {
             BaseArtifactType artifact = _srampClientAccessor.getClient().getArtifactMetaData(uuid);
             TargetBean bean = TargetFactory.toTarget(artifact);
@@ -92,15 +91,12 @@ public class TargetService implements ITargetService {
         }
     }
 
-    /*
-     * (non-Javadoc)
-     *
-     * @see
-     * org.overlord.dtgov.ui.client.shared.services.ITargetService#save(org.
-     * overlord.dtgov.ui.client.shared.beans.TargetBean)
+    /**
+     * @see org.overlord.dtgov.ui.client.shared.services.ITargetService#save(org.overlord.dtgov.ui.client.shared.beans.TargetBean)
      */
     @Override
     public String save(TargetBean target) throws DtgovUiException {
+        checkAuthorization();
         List<ValidationError> errors = _targetValidator.validate(target);
         if (errors.size() == 0) {
             BaseArtifactType artifact = TargetFactory.toBaseArtifact(target);
@@ -133,22 +129,19 @@ public class TargetService implements ITargetService {
 
     }
 
-    /*
-     * (non-Javadoc)
-     *
+    /**
      * @see org.overlord.dtgov.ui.client.shared.services.ITargetService#list()
      */
     @Override
     public List<TargetSummaryBean> list() throws DtgovUiException {
+        checkAuthorization();
 
         SrampAtomApiClient client = _srampClientAccessor.getClient();
 
         SrampClientQuery query = client.buildQuery("/s-ramp/ext/" + TargetConstants.TARGET_EXTENDED_TYPE); //$NON-NLS-1$
-
-        query = query.startIndex(0).orderBy("name"); //$NON-NLS-1$
-
+        query.startIndex(0).orderBy("name"); //$NON-NLS-1$
         query.propertyName(TargetConstants.TARGET_TYPE);
-        query = query.ascending();
+        query.ascending();
         try {
             QueryResultSet resultSet = query.query();
             List<TargetSummaryBean> summary = TargetFactory.asList(resultSet);
@@ -196,6 +189,17 @@ public class TargetService implements ITargetService {
      */
     public void setTargetValidator(TargetValidator targetValidator) {
         this._targetValidator = targetValidator;
+    }
+    
+    /**
+     * Checks that the current user is authorized to perform the action.
+     * @throws DtgovUiException
+     */
+    private static final void checkAuthorization() throws DtgovUiException {
+        ServletRequest request = RpcContext.getServletRequest();
+        if (!AuthUtils.isOverlordAdmin((HttpServletRequest) request)) {
+            throw new DtgovUiException(Messages.i18n.format("UserNotAuthorized")); //$NON-NLS-1$
+        }
     }
 
 }
